producer/requirements.txt
cffi==1.15.1
flake8==6.0.0
mccabe==0.7.0
pika==1.3.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python-dotenv==1.0.0
rabbitmq==0.2.0
wait-for-it==2.2.1